backend/src/requirements.txt
PyYAML==6.0
requests==2.22.0
SQLAlchemy==1.4.29
